% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gtfs_field_exists}
\alias{check_gtfs_field_exists}
\title{Check for field existence in a GTFS text file}
\usage{
check_gtfs_field_exists(gtfs, file, field)
}
\arguments{
\item{gtfs}{A GTFS object as created by \code{\link{read_gtfs}}.}

\item{file}{File in which the field should exist.}

\item{field}{Field to check existence of.}
}
\value{
If the check is successful, returns \code{TRUE} invisibly. Else,
  returns a string with error message.
}
\description{
Checks for the existence of a given field in a GTFS text file (represented by
a column in a \code{data.table}). Must be used in conjunction with
\code{\link[checkmate]{assert}}.
}
