% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gtfs_file_exists}
\alias{check_gtfs_file_exists}
\title{Check for GTFS text file existence}
\usage{
check_gtfs_file_exists(gtfs, file)
}
\arguments{
\item{gtfs}{A GTFS object as created by \code{\link{read_gtfs}}.}

\item{file}{File to check existence of.}
}
\value{
If the check is successful, returns \code{TRUE} invisibly. Else,
  returns a string with error message.
}
\description{
Checks for the existence of a given file (as represented by a
\code{data.table}) in a GTFS object. Must be used in conjunction with
\code{\link[checkmate]{assert}}.
}
