% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{convert_from_standard}
\alias{convert_from_standard}
\title{Convert a standards-compliant GTFS into a gtfstools' GTFS}
\usage{
convert_from_standard(gtfs)
}
\arguments{
\item{gtfs}{A GTFS object.}
}
\value{
The GTFS object passed to the \code{gtfs} parameter, after converting the
relevant fields.
}
\description{
Converts a standards-compliant GTFS into a gtfstools' GTFS (i.e. one in which
date fields are Date objects, not integers).
}
\keyword{internal}
