% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_stops_to_sf.R
\name{convert_stops_to_sf}
\alias{convert_stops_to_sf}
\title{Convert GTFS stops to simple feature object}
\usage{
convert_stops_to_sf(gtfs, crs = 4326)
}
\arguments{
\item{gtfs}{A GTFS object as created by \code{\link{read_gtfs}}.}

\item{crs}{The coordinate reference system (CRS) represented as an EPSG code.
Defaults to 4326 (latlong WGS84).}
}
\value{
A \code{POINT sf} object.
}
\description{
Returns the `stops.txt` data from a GTFS feed as a \code{POINT sf} object.
}
\examples{
# read gtfs
data_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")
gtfs <- read_gtfs(data_path)

stops_sf <- convert_stops_to_sf(gtfs)
head(stops_sf)


}
