% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{convert_to_standard}
\alias{convert_to_standard}
\title{Convert a gtfstools' GTFS into a standards-compliant GTFS}
\usage{
convert_to_standard(gtfs)
}
\arguments{
\item{gtfs}{A GTFS object, as created by \code{\link[=read_gtfs]{read_gtfs()}}.}
}
\value{
The GTFS object passed to the \code{gtfs} parameter, after converting the
relevant fields.
}
\description{
Converts a gtfstools' GTFS into a standards-compliant GTFS (i.e. date fields
are converted from Date to integer).
}
\keyword{internal}
