% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_gtfs_without_field}
\alias{copy_gtfs_without_field}
\title{Make a GTFS copy without a given field from a given file}
\usage{
copy_gtfs_without_field(gtfs, file, field)
}
\arguments{
\item{gtfs}{The GTFS to be copied, as created by \code{\link{read_gtfs}}.}

\item{file}{File from which the field must be removed.}

\item{field}{Field to be removed.}
}
\value{
A GTFS object without the given field.
}
\description{
Creates a copy of a GTFS object without a given field from a given file. Used
for testing.
}
