% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_gtfs_without_file}
\alias{copy_gtfs_without_file}
\title{Make a GTFS copy without a given file}
\usage{
copy_gtfs_without_file(gtfs, file)
}
\arguments{
\item{gtfs}{A GTFS object, as created by \code{\link[=read_gtfs]{read_gtfs()}}.}

\item{file}{File to be removed.}
}
\value{
A GTFS object without the given file.
}
\description{
Creates a copy of a GTFS object without a given file. Used for testing.
}
\keyword{internal}
