% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{date_to_integer}
\alias{date_to_integer}
\title{Convert a Date vector into an integer vector}
\usage{
date_to_integer(field)
}
\description{
Convert a Date vector into an integer vector
}
\keyword{internal}
