% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trip_length.R
\name{empty_trip_length}
\alias{empty_trip_length}
\title{Create an empty output of \code{get_trip_length()}}
\usage{
empty_trip_length(by)
}
\arguments{
\item{by}{Either \code{"trip"} or \code{"segment"}.}
}
\value{
An empty \code{data.table} with the columns of the output of
\code{get_trip_length()}.
}
\description{
Create an empty output of \code{get_trip_length()}
}
\keyword{internal}
