% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtfs.R
\name{extract_warning_message}
\alias{extract_warning_message}
\title{Extract warning message}
\usage{
extract_warning_message(warning_log)
}
\arguments{
\item{warning_log}{Warning log generated by \code{data.table::fread}.}
}
\value{
A string containing details on the parsing failure.
}
\description{
Extracts the warning message as a string from a warning log.
}
