% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_day_period.R
\name{filter_by_day_period}
\alias{filter_by_day_period}
\title{Filter GTFS object by day period}
\usage{
filter_by_day_period(
  gtfs,
  from,
  to,
  keep = TRUE,
  full_trips = TRUE,
  update_frequencies = FALSE
)
}
\arguments{
\item{gtfs}{A GTFS object.}

\item{from}{A string. The starting point of the day period, in the
"HH:MM:SS" format.}

\item{to}{A string. The ending point of the day period, in the "HH:MM:SS"
format.}

\item{keep}{A logical. Whether the entries related to the specified day
period should be kept or dropped (defaults to \code{TRUE}, which keeps the
entries).}

\item{full_trips}{A logical.}

\item{update_frequencies}{A logical.}
}
\value{
The GTFS object passed to the \code{gtfs} parameter, after the filtering
process.
}
\description{
Filters a GTFS object by a day period, keeping (or dropping) the relevant
entries in each file.
}
\examples{
data_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")
gtfs <- read_gtfs(data_path)

object.size(gtfs)

}
\seealso{
Other filtering functions: 
\code{\link{filter_by_agency_id}()},
\code{\link{filter_by_route_id}()},
\code{\link{filter_by_route_type}()},
\code{\link{filter_by_service_id}()},
\code{\link{filter_by_sf}()},
\code{\link{filter_by_shape_id}()},
\code{\link{filter_by_stop_id}()},
\code{\link{filter_by_trip_id}()},
\code{\link{filter_by_weekday}()}
}
\concept{filtering functions}
