% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_shape_id.R
\name{filter_shape_id}
\alias{filter_shape_id}
\title{Filter GTFS object by \code{shape_id}}
\usage{
filter_shape_id(gtfs, shape_id, keep = TRUE)
}
\arguments{
\item{gtfs}{A GTFS object.}

\item{shape_id}{A character vector. The \code{shape_id}s used to filter the data.}

\item{keep}{A logical. Whether the entries related to the passed \code{shape_id}s
should be kept or dropped (defaults to \code{TRUE}, which keeps the entries).}
}
\value{
The GTFS object passed to the \code{gtfs} parameter, after the filtering
process.
}
\description{
Filters a GTFS object by \code{shape_id}s, keeping (or dropping) the entries that
relate to such ids in each file.
}
\examples{
data_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")
gtfs <- read_gtfs(data_path)
shape_ids <- c("17846", "68962")

object.size(gtfs)

# keeps entries related to passed shape_ids
smaller_gtfs <- filter_shape_id(gtfs, shape_ids)
object.size(smaller_gtfs)

# drops entries related to passed shape_ids
smaller_gtfs <- filter_shape_id(gtfs, shape_ids, keep = FALSE)
object.size(smaller_gtfs)

}
\seealso{
Other filtering functions: 
\code{\link{filter_route_id}()},
\code{\link{filter_route_type}()},
\code{\link{filter_stop_id}()},
\code{\link{filter_trip_id}()}
}
\concept{filtering functions}
