% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_trip_id.R
\name{filter_trip_id}
\alias{filter_trip_id}
\title{Filter GTFS data by \code{trip_id}}
\usage{
filter_trip_id(gtfs, trip_id, keep = TRUE)
}
\arguments{
\item{gtfs}{A GTFS object.}

\item{trip_id}{A character vector. The \code{trip_id}s used to filter the data.}

\item{keep}{A logical. Whether the entries related to the passed \code{trip_id}s
should be kept or dropped (defaults to \code{TRUE}, which keeps the entries).}
}
\value{
The GTFS object passed to the \code{gtfs} parameter, after the filtering
process.
}
\description{
Filters a GTFS object by \code{trip_id}s, keeping (or dropping) the entries that
relate to such ids in each file.
}
\examples{
data_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")
gtfs <- read_gtfs(data_path)
trip_ids <- c("CPTM L07-0", "2002-10-0")

object.size(gtfs)

# keeps entries related to passed trip_ids
smaller_gtfs <- filter_trip_id(gtfs, trip_ids)
object.size(smaller_gtfs)

# drops entries related to passed trip_ids
smaller_gtfs <- filter_trip_id(gtfs, trip_ids, keep = FALSE)
object.size(smaller_gtfs)

}
