% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gtfs_meta}
\alias{get_gtfs_meta}
\title{Create GTFS metadata}
\usage{
get_gtfs_meta()
}
\value{
A GTFS metadata list.
}
\description{
Creates GTFS metadata (a list including all possible fields for each text
file, their column types and which ones are optional/required).
}
