% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{class}
\name{gtfs_date-class}
\alias{gtfs_date-class}
\title{Set class and method to read dates as formatted in GTFS to a 'Date' object}
\description{
This is a build-time dependency on methods, as opposed to a run-time
dependency, thus requiring the importFrom tag to avoid a NOTE when checking
the package on CRAN.
}
\keyword{internal}
