% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_isochrones.R
\name{gtfs_isochrones}
\alias{gtfs_isochrones}
\title{Produces isochrones as an sf-object}
\usage{
gtfs_isochrones(
  raptor_result,
  stops,
  breaks = c(3600, 2 * 3600, 3 * 3600),
  hull_alpha_min = 0.2,
  buffer_value = 4000
)
}
\arguments{
\item{raptor_result}{A result from tidytransit::raptor function}

\item{stops}{A stops dataframe obtained from a tidytransit gtfs object: gtfs$stops}

\item{breaks}{A vector containing the travel time breaks for the isochrone in seconds.}

\item{hull_alpha_min}{The alpha values increase by increasing break interval for better graphic display of the isochrones.}

\item{buffer_value}{The shapes are buffered in the end for graphic display. Value in meters.}
}
\value{
An sp object with features corresponding to the isochrones of the breaks provided in 'breaks'.
}
\description{
gtfs_isochrones Creates one SP Object containg a isochrones calculated from a tidytransit::raptor result.
The function uses the minimal travel times from the start stop to all stops as a main input parameter.
This isochrone is a not meant to depict exact reachibility of transit in detail. Use opentripplanner for this instead.
The goal here is to provide a decent overview of transit access at larger scales (regional, national or continental).
IMPORTANT: The raptor_result table has to have only have one single stop as origin
}
\examples{

}
