% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trip_length.R
\name{has_shapes}
\alias{has_shapes}
\title{Check whether a GTFS object has shapes linked to its trips}
\usage{
has_shapes(gtfs)
}
\arguments{
\item{gtfs}{A GTFS object, as created by \code{\link[=read_gtfs]{read_gtfs()}}.}
}
\value{
\code{TRUE} if the GTFS object has a \code{shapes} table and a \code{shape_id}
column in its \code{trips} table, \code{FALSE} otherwise.
}
\description{
Check whether a GTFS object has shapes linked to its trips
}
\keyword{internal}
