% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hms_to_string}
\alias{hms_to_string}
\title{Convert time to string}
\usage{
hms_to_string(time)
}
\arguments{
\item{time}{An \code{hms} object.}
}
\value{
A time-representing string..
}
\description{
Converts objects of class \code{hms} to strings in the "HH:MM:SS" format.
}
