% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{integer_to_date}
\alias{integer_to_date}
\title{Convert an integer vector into a Date vector}
\usage{
integer_to_date(field)
}
\description{
Convert an integer vector into a Date vector
}
\keyword{internal}
