% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_validator.R
\name{list_validator_versions}
\alias{list_validator_versions}
\title{List MobilityData's GTFS validator versions}
\usage{
list_validator_versions()
}
\value{
A data.table with each CLI version, published date, and download URL.
}
\description{
Lists available versions of MobilityData's command line tool to validate GTFS feeds.
}
\seealso{
Other validation:
\code{\link[=download_validator]{download_validator()}},
\code{\link[=validate_gtfs]{validate_gtfs()}}
}
\concept{validation}
