% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trip_length.R
\name{map_deprecated_file}
\alias{map_deprecated_file}
\title{Map the deprecated \code{file} argument to \code{method}}
\usage{
map_deprecated_file(file, fn_name)
}
\arguments{
\item{file}{The value given to the deprecated \code{file} argument.}

\item{fn_name}{The name of the function whose argument is deprecated.}
}
\value{
Either \code{"shapes"}, if \code{"shapes"} is included in \code{file}, or
\code{"euclidean"}.
}
\description{
Raises a deprecation warning and returns the \code{method} that corresponds to
the given \code{file}.
}
\keyword{internal}
