% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gtfs.R
\name{merge_gtfs}
\alias{merge_gtfs}
\title{Merge GTFS files}
\usage{
merge_gtfs(..., files = NULL, warnings = TRUE)
}
\arguments{
\item{...}{GTFS objects to be merged. Each argument can either be a GTFS or
a list of GTFS objects.}

\item{files}{A character vector listing the GTFS tables to be merged. If
\code{NULL} (the default), all tables are merged.}

\item{warnings}{Whether to display warning messages (defaults to \code{TRUE}).}
}
\value{
A GTFS object in which each table is a combination (by row) of the
tables from the specified GTFS objects.
}
\description{
Combines many GTFS objects into a single one.
}
\examples{
spo_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")
ggl_path <- system.file("extdata/ggl_gtfs.zip", package = "gtfstools")

spo_gtfs <- read_gtfs(spo_path)
names(spo_gtfs)

ggl_gtfs <- read_gtfs(ggl_path)
names(ggl_gtfs)

merged_gtfs <- merge_gtfs(spo_gtfs, ggl_gtfs)
names(merged_gtfs)

# use a list() to programatically merge many GTFS objects
merged_gtfs <- merge_gtfs(list(spo_gtfs, ggl_gtfs))

}
