% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_date.R
\name{parse_dates}
\alias{parse_dates}
\title{Parse dates given as \code{Date}s or strings}
\usage{
parse_dates(date, call = parent.frame())
}
\arguments{
\item{date}{The \code{date} argument to be parsed.}

\item{call}{The environment of the function whose \code{date} argument is
parsed, in which errors are reported.}
}
\value{
A \code{Date} vector.
}
\description{
Validates a \code{date} argument given either as a \code{Date} vector or as a
character vector of dates in the \code{"YYYYMMDD"} or \code{"YYYY-MM-DD"} formats
(which may be mixed), and converts it to a \code{Date} vector.
}
\keyword{internal}
