% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_sf.R
\name{polygon_from_bbox}
\alias{polygon_from_bbox}
\title{Build a bouding box polygon}
\usage{
polygon_from_bbox(bbox)
}
\arguments{
\item{bbox}{A bouding box object.}
}
\description{
Builds a polygon from a bounding box object.
}
\keyword{internal}
