% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtfs.R
\name{read_gtfs}
\alias{read_gtfs}
\title{Read and validate GTFS files}
\usage{
read_gtfs(path, files = NULL, quiet = FALSE)
}
\arguments{
\item{path}{The path to a GTFS \code{.zip} file.}

\item{files}{A character vector containing the text files to be read from the
GTFS (without the \code{.txt} extension). If \code{NULL} (the default) all
existing files are read.}

\item{quiet}{Whether to display warning messages (the default) or not.}
}
\value{
A GTFS object: a list of data.tables in which each index represents a
  GTFS text file.
}
\description{
Reads GTFS text files from a zip file and validates them against GTFS
specifications.
}
