% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_gtfs_by_route_type.R
\name{reduce_gtfs_to_rail}
\alias{reduce_gtfs_to_rail}
\title{Reduce a GTFS file to a subset based on route_type}
\usage{
reduce_gtfs_to_rail(gtfs_in, route_type)
}
\arguments{
\item{gtfs_in}{tidytransit GTFS object}

\item{numeric}{route_type to reduce the GTFS data to}
}
\value{
A tidytransit GTFS object filtered by the selected route type
}
\description{
Reduce a GTFS file to a subset based on route_type
}
\examples{

}
