% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seconds_to_string}
\alias{seconds_to_string}
\title{Convert seconds after midnight to time string}
\usage{
seconds_to_string(seconds)
}
\arguments{
\item{seconds}{An integer.}
}
\value{
A time-representing string.
}
\description{
Converts seconds after midnight as integers to strings in the "HH:MM:SS"
format.
}
\keyword{internal}
