% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_to_stop_duration.R
\name{stop_to_stop_duration}
\alias{stop_to_stop_duration}
\title{Return the duration of segments between stops}
\usage{
stop_to_stop_duration(gtfs, trip_id = NULL)
}
\arguments{
\item{gtfs}{A GTFS file as read by \code{tidytransit::read_gtfs()}.}

\item{trip_id}{A string vector including the \code{trip_id}s to have their
segments' duration calculated. If \code{NULL} (the default) the function
calculates the segments' duration for every \code{trip_id} in the GTFS.}
}
\value{
A table containing the duration of the segments of each specified
  trip.
}
\description{
Returns the duration of segments between stops of each specified
\code{trip_id}.
}
