% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_to_stop_length.R
\name{stop_to_stop_length}
\alias{stop_to_stop_length}
\title{Return the length of segments between stops}
\usage{
stop_to_stop_length(
  gtfs,
  trip_id = NULL,
  file = "stop_times",
  drop_geometry = TRUE,
  crs = 4326
)
}
\arguments{
\item{gtfs}{A GTFS file as read by \code{tidytransit::read_gtfs()}.}

\item{trip_id}{A string vector including the \code{trip_id}s to have their
segments' length calculated. If \code{NULL} (the default) the function
calculates the segments' length for every \code{trip_id} in the GTFS.}

\item{file}{The file from which geometries should be generated. By now
supports only \code{stop_times}.}

\item{drop_geometry}{Whether the function should drop segment's geometry
(defaults to \code{TRUE}).}

\item{crs}{The CRS of the resulting object. Defaults to 4326 (WGS 84).}
}
\value{
A table containing the segments of each specified trip, their length
  (in meters) and their geometry, if desired.
}
\description{
Returns the length of segments between stops of each specified \code{trip_id},
based either on the \code{shapes} (not yet supported) or the
\code{stop_times}.
}
