% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_to_hms}
\alias{string_to_hms}
\title{Convert string to time}
\usage{
string_to_hms(string)
}
\arguments{
\item{string}{A string in "HH:MM:SS" format.}
}
\value{
An \code{hms} object.
}
\description{
Converts strings in the "HH:MM:SS" format to \code{hms}.
}
