% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_to_seconds}
\alias{string_to_seconds}
\title{Convert time string to seconds after midnight}
\usage{
string_to_seconds(string)
}
\arguments{
\item{string}{A string in "HH:MM:SS" format.}
}
\value{
The seconds after midnight of a given time string as an integer.
}
\description{
Converts strings in the "HH:MM:SS" format to seconds after midnight.
}
\keyword{internal}
