% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_field_exists}
\alias{test_field_exists}
\title{Test for field existence in a GTFS text file}
\usage{
test_field_exists(gtfs, file, field)
}
\arguments{
\item{gtfs}{A GTFS object as created by \code{\link{read_gtfs}}.}

\item{file}{File in which the field should exist.}

\item{field}{Field to check existence of.}
}
\value{
\code{TRUE}, if the check is successful, and \code{FALSE} otherwise.
}
\description{
Test for the existence of a given field in a GTFS text file (represented by
a column in a \code{data.table}).
}
\keyword{internal}
