% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_file_exists}
\alias{test_file_exists}
\title{Test for GTFS text file existence}
\usage{
test_file_exists(gtfs, file)
}
\arguments{
\item{gtfs}{A GTFS object as created by \code{\link{read_gtfs}}.}

\item{file}{File to test the existence of.}
}
\value{
\code{TRUE}, if the check is successful, and \code{FALSE} otherwise.
}
\description{
Tests for the existence of a given file (as represented by a
\code{data.table}) in a GTFS object.
}
\keyword{internal}
