% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_duration.R
\name{trip_duration}
\alias{trip_duration}
\title{Return trip duration}
\usage{
trip_duration(gtfs, trip_id = NULL)
}
\arguments{
\item{gtfs}{A GTFS file as read by \code{tidytransit::read_gtfs()}.}

\item{trip_id}{A string vector including the \code{trip_id}s to have their
durations calculated. If \code{NULL} (the default) the function calculates
the duration of every \code{trip_id} in the GTFS.}
}
\value{
A table containing the duration (in minutes) of each specified trip.
}
\description{
Returns the duration of each specified \code{trip_id}.
}
