% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips_geometries.R
\name{trips_geometries}
\alias{trips_geometries}
\title{Return trips geometries}
\usage{
trips_geometries(
  gtfs,
  trip_id = NULL,
  file = c("shapes", "stop_times"),
  crs = 4326
)
}
\arguments{
\item{gtfs}{A \code{gtfs} file as read by \code{tidytransit::read_gtfs()}.}

\item{trip_id}{A string vector including the \code{trip_id}s to have their
geometries generated. If \code{NULL} (the default) the function generates
geometries for every \code{trip_id} in the \code{gtfs}.}

\item{file}{The file from which geometries should be generated. By default
uses both \code{shapes} and \code{stop_times}.}

\item{crs}{The CRS of the resulting object. Defaults to 4326 (WGS 84).}
}
\value{
A \code{LINESTRING sf}.
}
\description{
Returns each \code{trip_id} geometry, based either on the \code{shapes} or
the \code{stop_times} file (or both).
}
