% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{warn_missing_ids}
\alias{warn_missing_ids}
\title{Warn about ids missing from a table}
\usage{
warn_missing_ids(id, existing_id, table, id_name)
}
\arguments{
\item{id}{A character vector of ids given by the user.}

\item{existing_id}{A character vector of the ids that exist in the table.}

\item{table}{The name of the table, used in the warning message.}

\item{id_name}{The name of the id field, used in the warning message.}
}
\value{
The missing ids, invisibly.
}
\description{
Raises a warning listing the elements of \code{id} not found in \code{existing_id}.
The warning is attributed to the function that called this helper.
}
\keyword{internal}
