% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_stop_times.R
\name{warn_times_not_interpolated}
\alias{warn_times_not_interpolated}
\title{Warn about stop times that couldn't be interpolated}
\usage{
warn_times_not_interpolated(n_left, left_trips, back_trips)
}
\arguments{
\item{n_left}{The number of stop times left blank.}

\item{left_trips}{The trips with blank stop times outside two timepoints or
whose distances couldn't be calculated.}

\item{back_trips}{The trips with blank stop times between timepoints in
which the arrival is earlier than the previous departure.}
}
\value{
Invisibly returns \code{NULL}. Called for its warning.
}
\description{
Warn about stop times that couldn't be interpolated
}
\keyword{internal}
