% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_sf.R
\name{filter_by_sf}
\alias{filter_by_sf}
\title{Filter a GTFS object using a \verb{simple features} object (defunct)}
\usage{
filter_by_sf(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
Doesn't return: always raises an error.
}
\description{
This function was deprecated in gtfstools 1.3.0 and is now defunct: calling
it raises an error. Please use \code{\link[=filter_by_spatial_extent]{filter_by_spatial_extent()}} instead, which
takes the same arguments.
}
\seealso{
\code{\link[=filter_by_spatial_extent]{filter_by_spatial_extent()}}, which replaces this function.
}
