% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_sf.R
\name{filter_by_sf}
\alias{filter_by_sf}
\title{Filter a GTFS object using a \verb{simple features} object}
\usage{
filter_by_sf(gtfs, geom, spatial_operation = sf::st_intersects, keep = TRUE)
}
\arguments{
\item{gtfs}{A GTFS object.}

\item{geom}{An \code{sf} object. Describes the geometry used to filter the data.}

\item{spatial_operation}{An spatial operation function from the set of
options listed in \link[sf:geos_binary_pred]{geos_binary_pred}. Defaults to
\code{sf::st_intersects}, which tests if the shapes and trips have ANY
intersection with the object specified in \code{geom}.}

\item{keep}{A logical. Whether the entries related to the shapes and trips
that cross through the given geometry should be kept or dropped (defaults
to \code{TRUE}, which keeps the entries).}
}
\value{
The GTFS object passed to the \code{gtfs} parameter, after the filtering
process.
}
\description{
Filters a GTFS object using the geometry of an \code{sf} object, keeping (or
dropping) entries related to shapes and trips selected through an spatial
operation.
}
\examples{
data_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")
gtfs <- read_gtfs(data_path)

shape_id <- "68962"
shape_sf <- convert_shapes_to_sf(gtfs, shape_id)
bbox <- sf::st_bbox(shape_sf)
object.size(gtfs)

polygon <- sf::st_polygon(
  list(rbind(
    c(bbox$xmin, bbox$ymin),
    c(bbox$xmin, bbox$ymax),
    c(bbox$xmax, bbox$ymax),
    c(bbox$xmax, bbox$ymin),
    c(bbox$xmin, bbox$ymin)
  ))
)
polygon <- sf::st_sf(geom = sf::st_sfc(polygon), crs = 4326)

# keeps entries that intersect with the specified polygon
smaller_gtfs <- filter_by_sf(gtfs, polygon)
object.size(smaller_gtfs)

# drops entries that intersect with the specified polygon
smaller_gtfs <- filter_by_sf(gtfs, polygon, keep = FALSE)
object.size(smaller_gtfs)

}
\seealso{
Other filtering functions: 
\code{\link{filter_by_route_id}()},
\code{\link{filter_by_route_type}()},
\code{\link{filter_by_shape_id}()},
\code{\link{filter_by_stop_id}()},
\code{\link{filter_by_trip_id}()}
}
\concept{filtering functions}
