% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_sf.R
\name{filter_by_sf}
\alias{filter_by_sf}
\title{Filter a GTFS object using a \verb{simple features} object (deprecated)}
\usage{
filter_by_sf(gtfs, geom, spatial_operation = sf::st_intersects, keep = TRUE)
}
\arguments{
\item{gtfs}{A GTFS object, as created by \code{\link[=read_gtfs]{read_gtfs()}}.}

\item{geom}{An \code{sf} object. Describes the geometry used to filter the data.}

\item{spatial_operation}{A spatial operation function from the set of
options listed in \link[sf:geos_binary_pred]{geos_binary_pred} (check the
\href{https://en.wikipedia.org/wiki/DE-9IM}{DE-I9M} Wikipedia entry for the
definition of each function). Defaults to \code{sf::st_intersects}, which tests
if the shapes and trips have ANY intersection with the object specified in
\code{geom}. Please note that \code{geom} is passed as the \code{x} argument of these
functions.}

\item{keep}{A logical. Whether the entries related to the shapes and trips
that cross through the given geometry should be kept or dropped (defaults
to \code{TRUE}, which keeps the entries).}
}
\value{
The GTFS object passed to the \code{gtfs} parameter, after the filtering
process.
}
\description{
This function has been deprecated as of the current package version and will
be completely removed from version 2.0.0 onward. Please use
\code{\link[=filter_by_spatial_extent]{filter_by_spatial_extent()}} instead.

Filters a GTFS object using the geometry of an \code{sf} object, keeping (or
dropping) entries related to shapes and trips selected through a spatial
operation.
}
\examples{
\dontshow{
  old_dt_threads <- data.table::setDTthreads(1)
  on.exit(data.table::setDTthreads(old_dt_threads), add = TRUE)
}
data_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")
gtfs <- read_gtfs(data_path)

shape_id <- "68962"
shape_sf <- convert_shapes_to_sf(gtfs, shape_id)
bbox <- sf::st_bbox(shape_sf)
object.size(gtfs)

# keeps entries that intersect with the specified polygon
smaller_gtfs <- filter_by_sf(gtfs, bbox)
object.size(smaller_gtfs)

# drops entries that intersect with the specified polygon
smaller_gtfs <- filter_by_sf(gtfs, bbox, keep = FALSE)
object.size(smaller_gtfs)

# uses a different function to filter the gtfs
smaller_gtfs <- filter_by_sf(gtfs, bbox, spatial_operation = sf::st_contains)
object.size(smaller_gtfs)

}
\seealso{
Other filtering functions: 
\code{\link{filter_by_agency_id}()},
\code{\link{filter_by_route_id}()},
\code{\link{filter_by_route_type}()},
\code{\link{filter_by_service_id}()},
\code{\link{filter_by_shape_id}()},
\code{\link{filter_by_spatial_extent}()},
\code{\link{filter_by_stop_id}()},
\code{\link{filter_by_time_of_day}()},
\code{\link{filter_by_trip_id}()},
\code{\link{filter_by_weekday}()}
}
\concept{filtering functions}
