% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_route_id.R
\name{filter_route_id}
\alias{filter_route_id}
\title{Filter GTFS data by route ids}
\usage{
filter_route_id(gtfs, route_ids, keep = TRUE)
}
\arguments{
\item{gtfs}{A GTFS object as created by \code{\link{read_gtfs}}.}

\item{route_ids}{A vector of route ids belonging to the routes of the
gtfs data. A `route_id`can be either a string or a number,
depending on the GTF input.}

\item{keep}{Logical. Whether the passed `route_ids` should be kept (Defaults
to `TRUE`), or whether they shoud be droped.}
}
\value{
A filtered GTFS object
}
\description{
Filter a GTFS data by its route ids, subsetting routes
and trips. It also removes the unnecessary stop_times, shapes, frequencies
(if exist in a feed), and stops accordingly.
}
\examples{
# read gtfs
data_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")
gtfs <- read_gtfs(data_path)

# filter: Keep selected routes
subset1 <- filter_route_id(gtfs, route_ids=c("6450-51", "CPTM L11"), keep = T)

# filter: Drop selected routes
subset2 <- filter_route_id(gtfs, route_ids=c("6450-51", "CPTM L11"), keep = F)

}
