% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_route_type.R
\name{filter_route_type}
\alias{filter_route_type}
\title{Filter GTFS data by route type (transport mode)}
\usage{
filter_route_type(gtfs, route_types, keep = TRUE)
}
\arguments{
\item{gtfs}{A GTFS object as created by \code{\link{read_gtfs}}.}

\item{route_types}{A vector of route types belonging to the routes of the
GTFS data. Each transport mode is associated with a
`route_type` number. See Details.}

\item{keep}{Logical. Whether the passed `route_ids` should be kept (Defaults
to `TRUE`), or whether they should be dropped.}
}
\value{
Returns a GTFS object with filtered routes, trips, shapes, stops,
        stop_times and frequencies.
}
\description{
Filter a GTFS data byits route type (transport mode), subsetting routes
and trips. It also removes the unnecessary stop_times, shapes, frequencies
(if exist in a feed), and stops accordingly.
}
\details{
Route types:
|**`route_type`**|**Transport mode**| Note |
|-----|-----|-----|
|0  | Tram, Streetcar, Light rail | Any light rail or street level system within a metropolitan area |
|1  | Subway, Metro | Any underground rail system within a metropolitan area |
|2  | Rail | Used for intercity or long-distance travel |
|3  | Bus | Used for short- and long-distance bus routes |
|4  | Ferry | Used for short- and long-distance boat service |
|5  | Cable tram | Used for street-level rail cars where the cable runs beneath the vehicle, e.g., cable car in San Francisco |
|6  | Aerial lift, suspended cable car (e.g., gondola lift, aerial tramway) | Cable transport where cabins, cars, gondolas or open chairs are suspended by means of one or more cables |
|7  | Funicular | Any rail system designed for steep inclines |
|11 |  Trolleybus | Electric buses that draw power from overhead wires using poles |
|12 |  Monorail | Railway in which the track consists of a single rail or a beam |
}
\examples{
# read gtfs
data_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")
gtfs <- read_gtfs(data_path)

# filter: Keep selected route types
subset1 <- filter_route_type(gtfs, route_types=c(2, 3), keep = TRUE)

# filter: Drop selected route types
subset2 <- filter_route_type(gtfs, route_types=c(2, 3), keep = FALSE)

}
\concept{filter functions}
