% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies_to_stop_times.R
\name{frequencies_to_stop_times}
\alias{frequencies_to_stop_times}
\title{Convert frequencies to stop times}
\usage{
frequencies_to_stop_times(gtfs, trip_id = NULL)
}
\arguments{
\item{trip_id}{}
}
\value{

}
\description{
Convert frequencies to stop times
}
\examples{
data_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")
gtfs <- read_gtfs(data_path)


}
