% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_route_frequency.R
\name{get_route_frequency}
\alias{get_route_frequency}
\title{Get route frequency}
\usage{
get_route_frequency(gtfs, route_id = NULL, from, to)
}
\arguments{
\item{gtfs}{A GTFS object, as created by \code{\link[=read_gtfs]{read_gtfs()}}.}

\item{route_id}{A character vector including the \code{route_id}s to have their
frequencies calculated. If \code{NULL} (the default), the function calculates
the speed of every \code{route_id} in the GTFS.}

\item{from}{A string. The starting point of the time of day, in the
"HH:MM:SS" format.}

\item{to}{A string. The ending point of the time of day, in the "HH:MM:SS"
format.}
}
\value{
A \code{data.table} containing the headway (in seconds) and number of
departures of each specified \code{route_id} within a given time period.
}
\description{
Returns the headway (in seconds) and the number of departures of each
specified \code{route_id} within a given time period.
}
\examples{
data_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")

gtfs <- read_gtfs(data_path)

# all routes
route_frequency <- get_route_frequency(gtfs,
                                       from = '08:00:00',
                                       to = '20:00:00')
head(route_frequency)

# selected routes
route_ids <- c("CPTM L07", "2002-10")
route_frequency <- get_route_frequency(gtfs,
                                       route_id = route_ids,
                                       from = '08:00:00',
                                       to = '20:00:00')
print(route_frequency)

}
\concept{frequency functions}
