% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_date.R
\name{get_services_on_days}
\alias{get_services_on_days}
\title{Get the services that run on the given days}
\usage{
get_services_on_days(gtfs, days)
}
\arguments{
\item{gtfs}{A GTFS object, as created by \code{\link[=read_gtfs]{read_gtfs()}}.}

\item{days}{An integer vector of unique days, as the number of days since
1970-01-01.}
}
\value{
A character vector of unique \code{service_id}s, empty if no service
runs on the given days.
}
\description{
Returns the services that run on at least one of the given days, according
to the \code{calendar} and \code{calendar_dates} tables (either may be missing).
}
\keyword{internal}
