% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trip_geometry.R
\name{get_trip_geometry}
\alias{get_trip_geometry}
\title{Get trip geometry}
\usage{
get_trip_geometry(
  gtfs,
  trip_id = NULL,
  file = c("shapes", "stop_times"),
  crs = 4326
)
}
\arguments{
\item{gtfs}{A GTFS object as created by \code{\link{read_gtfs}}.}

\item{trip_id}{A string vector including the \code{trip_id}s to have their
geometries generated. If \code{NULL} (the default), the function generates
geometries for every \code{trip_id} in the GTFS.}

\item{file}{The file from which geometries should be generated. By default
uses both \code{shapes} and \code{stop_times}.}

\item{crs}{The CRS of the resulting object. Defaults to 4326 (WGS 84).}
}
\value{
A \code{LINESTRING sf}.
}
\description{
Returns the geometry of each specified \code{trip_id}, based either on the
\code{shapes} or the \code{stop_times} file (or both).
}
\section{Details}{

The geometry generation works differently for the two files. In the case of
\code{shapes}, the shape as described in the text file is converted to an
\code{sf} object. For \code{stop_times} the geometry is the result of linking
subsequent stops along a straight line (stops' coordinates are retrieved from
the \code{stops} file). Thus, the resolution of the geometry when generated
with \code{shapes} tends to be much higher than when created with
\code{stop_times}.
}

