% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfstools.R
\docType{package}
\name{gtfstools}
\alias{gtfstools}
\alias{gtfstools-package}
\title{gtfstools: General Transit Feed Specification (GTFS) Editing and Analysing
Tools}
\description{
Utility functions to read, manipulate, analyse and write transit feeds in the
General Transit Feed Specification (GTFS) data format.
}
\section{Usage}{

Please check the vignettes for more on the package usage:
\itemize{
\item Basic usage: reading, analysing, manipulating and writing feeds. Run
\code{vignette("gtfstools")} or check it on the \href{https://ipea.github.io/gtfstools/articles/gtfstools.html}{website}.
\item Filtering GTFS feeds. Run \code{vignette("filtering", package = "gtfstools")} or
check it on the \href{https://ipea.github.io/gtfstools/articles/filtering.html}{website}.
\item Validating GTFS feeds. Run \code{vignette("validating", package = "gtfstools")}
or check it on the
\href{https://ipea.github.io/gtfstools/articles/validating.html}{website}.
\item Checking the calendar overlap of GTFS feeds. Run
\code{vignette("calendar_overlap", package = "gtfstools")} or check it on the
\href{https://ipea.github.io/gtfstools/articles/calendar_overlap.html}{website}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://ipea.github.io/gtfstools/}
  \item \url{https://github.com/ipea/gtfstools}
  \item Report bugs at \url{https://github.com/ipea/gtfstools/issues}
}

}
\author{
\strong{Maintainer}: Daniel Herszenhut \email{dhersz@gmail.com} (\href{https://orcid.org/0000-0001-8066-1105}{ORCID})

Authors:
\itemize{
  \item Daniel Herszenhut \email{dhersz@gmail.com} (\href{https://orcid.org/0000-0001-8066-1105}{ORCID})
  \item Rafael H. M. Pereira (\href{https://orcid.org/0000-0003-2125-7465}{ORCID})
  \item Pedro R. Andrade (\href{https://orcid.org/0000-0001-8675-4046}{ORCID})
  \item Joao Bazzo (\href{https://orcid.org/0000-0003-4536-5006}{ORCID})
}

Other contributors:
\itemize{
  \item Mark Padgham [contributor]
  \item Christopher Higgins [contributor]
  \item Arthur Bazolli \email{baz.arthur@gmail.com} [contributor]
  \item Marcus Saraiva (\href{https://orcid.org/0000-0001-6218-2338}{ORCID}) [contributor]
  \item Ipea - Institute for Applied Economic Research [copyright holder, funder]
}

}
\keyword{internal}
