% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfstools.R
\docType{package}
\name{gtfstools}
\alias{gtfstools}
\alias{_PACKAGE}
\alias{gtfstools-package}
\title{gtfstools: General Transit Feed Specification (GTFS) Editing and Analysing
Tools}
\description{
Utility functions to read, manipulate, analyse and write transit feeds in the
General Transit Feed Specification (GTFS) data format.
}
\section{Basic usage}{

Please check the introductory vignette for basic usage:
\itemize{
  \item{Run \code{vignette("gtfstools")};}
  \item{Alternatively, check it on
    \href{https://ipeagit.github.io/gtfstools/articles/gtfstools.html}{
    \code{{gtfstools}} website}.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://ipeagit.github.io/gtfstools/}
  \item \url{https://github.com/ipeaGIT/gtfstools}
  \item Report bugs at \url{https://github.com/ipeaGIT/gtfstools/issues}
}

}
\author{
\strong{Maintainer}: Daniel Herszenhut \email{dhersz@gmail.com} (\href{https://orcid.org/0000-0001-8066-1105}{ORCID})

Other contributors:
\itemize{
  \item Ipea - Institute for Applied Economic Research [copyright holder, funder]
}

}
