% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtfs.R
\name{read_files}
\alias{read_files}
\title{Read GTFS text files}
\usage{
read_files(file, temp_dir)
}
\arguments{
\item{file}{The name of the file (with \code{.txt} extension) to be read.}

\item{temp_dir}{The path to the temporary folder where the GTFS was unzipped.}
}
\value{
Either a data.table containing the desired file or a log message if
  any parsing warnings were thrown.
}
\description{
Reads a text file from the main zip file.
}
