% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_trip_speed.R
\name{set_trip_speed}
\alias{set_trip_speed}
\title{Set trip average speed}
\usage{
set_trip_speed(gtfs, trip_id, speed, unit = "km/h", by_reference = FALSE)
}
\arguments{
\item{gtfs}{A GTFS object as created by \code{\link{read_gtfs}}.}

\item{trip_id}{A string vector including the \code{trip_id}s to have their
average speed set.}

\item{speed}{A numeric representing the speed to be set. Its length either
equals 1, in which case the value is recycled for all \code{trip_id}s, or
equals \code{trip_id}'s length.}

\item{unit}{A string representing the unit in which the speed is given. One
of \code{"km/h"} (the default) or \code{"m/s"}.}

\item{by_reference}{Whether to update \code{stop_times}' \code{data.table} by
reference. Defaults to \code{FALSE}.}
}
\value{
If \code{by_reference} is set to \code{FALSE}, returns a GTFS object
  with the time columns of its \code{stop_times} adjusted. Else, returns a
  GTFS object invisibly (note that in this case the original GTFS object is
  altered).
}
\description{
Sets the average speed of each specified \code{trip_id} by changing the
\code{arrival_time} and \code{departure_time} columns in \code{stop_times}.
}
