% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gtfs.R
\name{validate_gtfs}
\alias{validate_gtfs}
\title{Validate GTFS file}
\usage{
validate_gtfs(gtfs, files, quiet)
}
\arguments{
\item{gtfs}{A GTFS object as created by \code{read_gtfs}.}

\item{files}{A character vector containing the text files to be validated
against the GTFS specification (without the \code{.txt} extension). If
\code{NULL} (the default) the provided GTFS is validated against all
possible GTFS text files.}

\item{quiet}{Whether to display warning messages (the default) or not.}
}
\value{
A data.table containing the validation summary of all possible fields
  from the specified files.
}
\description{
Validates the GTFS file against GTFS specifications and raises warnings if
required files/fields are not found.
}
