% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gtfs.R
\name{validate_gtfs}
\alias{validate_gtfs}
\title{Validate GTFS feed}
\usage{
validate_gtfs(gtfs, output_path, validator_path)
}
\arguments{
\item{gtfs}{The path to the GTFS feed to be validated.}

\item{output_path}{The path to the directory that the validator will create
and in which the results will be saved to.}

\item{validator_path}{The path to the GTFS validator, previously downloaded
with \code{\link[=download_validator]{download_validator()}}.}
}
\value{
Invisibly returns the normalized path to the directory where the
validation results were saved to.
}
\description{
Uses MobilityData's \href{https://github.com/MobilityData/gtfs-validator}{GTFS validator} to perform a GTFS
business rule validation. The results are available as an HTML report and in
JSON format. Please check the complete set of rules used in the validation
\href{https://github.com/MobilityData/gtfs-validator/blob/master/RULES.md}{here}.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
gtfs_path <- system.file("extdata/spo_gtfs.zip", package = "gtfstools")
output_path <- tempfile("validation_result")
validator_path <- download_validator(tempdir())

validate_gtfs(gtfs_path, output_path, validator_path)
\dontshow{\}) # examplesIf}
}
\seealso{
Other validation: 
\code{\link{download_validator}()}
}
\concept{validation}
