% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gtfs.R
\name{write_gtfs}
\alias{write_gtfs}
\title{Write GTFS files}
\usage{
write_gtfs(gtfs, path, optional = TRUE, extra = TRUE, overwrite = TRUE)
}
\arguments{
\item{gtfs}{A GTFS object as created by \code{read_gtfs}.}

\item{path}{The path to the \code{.zip} file in which the feed should be
written to.}

\item{optional}{Whether to write optional \code{.txt}. Defaults to TRUE.}

\item{extra}{Whether to write extra \code{.txt}. Defaults to TRUE.}

\item{overwrite}{Whether to overwrite existing \code{.zip} file. Defaults to
TRUE.}
}
\value{
Invisibly returns the provided GTFS object.
}
\description{
Writes in-memory GTFS objects as GTFS \code{.zip} files. Conditionally
includes optional and extra \code{.txt} files (check \code{\link{validate_gtfs}}
documentation to check what are optional/extra files).
}
\seealso{
\code{\link{validate_gtfs}}
}
